+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 Improvements:
 * [OE-86](https://openlmis.atlassian.net/browse/OE-86): GET /facilities/{id}/approvedProducts supports multiple program id
  query parameters
+* [OE-86](https://openlmis.atlassian.net/browse/OE-86): GET /programs has new program code repeatable query parameter
 
 Bux fixes:
 * [OLMIS-8068](https://openlmis.atlassian.net/browse/OLMIS-8068): Fix NPE on adding programs to products